requirements.txt
absl-py==0.9.0
amqp==2.5.2
astor==0.8.1
beautifulsoup4==4.8.2
billiard==3.6.2.0
bs4==0.0.1
cachetools==4.0.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SocketIO==4.2.1
gast==0.2.2
get==2019.4.13
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kombu==4.6.7
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
monotonic==1.5
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
post==2019.4.13
protobuf==3.11.3
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.10.1
python-dotenv==0.10.5
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
query-string==2019.4.13
redis==3.4.1
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
soupsieve==1.9.5
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.8
vine==1.3.0
Werkzeug==0.16.1
wrapt==1.12.0
